+++ Mapper/Fieldset.php
@@ -10,6 +10,7 @@
 
 namespace Austral\FormBundle\Mapper;
 
+use Austral\EntityBundle\Entity\EntityInterface;
 use Austral\FormBundle\Mapper\Base\MapperElementInterface;
 use Austral\FormBundle\Mapper\Base\MapperElement;
 
@@ -39,33 +40,57 @@ class Fieldset extends MapperElement implements MapperElementInterface
   protected bool $viewName = true;
 
   /**
+   * @var string|null
+   */
+  protected ?string $entitled = null;
+
+  /**
    * @var array
    */
   protected array $attr = array();
+
+  /**
+   * @var bool
+   */
+  protected bool $collapse = false;
+
+  /**
+   * @var array
+   */
+  protected array $translateArguments = array();
+
+  /**
+   * @var \Closure|null
+   */
+  protected ?\Closure $closureTranslateArgument = null;
 
   /**
    * Choices constructor.
    *
    * @param string $name
    * @param MapperElementInterface $parent
+   * @param string|null $entitled
    */
-  public function __construct(string $name, MapperElementInterface $parent)
+  public function __construct(string $name, MapperElementInterface $parent, ?string $entitled = null)
   {
     $this->name = $name;
+    $this->entitled = $entitled ?? $name;
     $slugger = new AsciiSlugger();
     $this->keyname = "fieldset-".strtolower($slugger->slug($name));
     $this->parent = $parent;
   }
 
   /**
+   * @param string|null $afterId
+   *
    * @return string
    */
-  public function getFieldsetId(): string
+  public function getFieldsetId(string $afterId = null): string
   {
     if($this->positionName == self::POSITION_MASTER)
     {
       $slugger = new AsciiSlugger();
-      $fieldsetId = "fieldset-".strtolower($slugger->slug($this->name));
+      $fieldsetId = "fieldset-".strtolower($slugger->slug($this->name)).($afterId ? "-{$afterId}" : null);
     }
     else
     {
@@ -91,6 +116,25 @@ class Fieldset extends MapperElement implements MapperElementInterface
   public function getPositionName(): string
   {
     return $this->positionName;
+  }
+
+  /**
+   * @return string|null
+   */
+  public function getEntitled(): ?string
+  {
+    return $this->entitled;
+  }
+
+  /**
+   * @param string|null $entitled
+   *
+   * @return Fieldset
+   */
+  public function setEntitled(?string $entitled): Fieldset
+  {
+    $this->entitled = $entitled;
+    return $this;
   }
 
   /**
@@ -132,11 +176,78 @@ class Fieldset extends MapperElement implements MapperElementInterface
   }
 
   /**
+   * @param EntityInterface|null $object
+   *
+   * @return array
+   */
+  public function getTranslateArguments(?EntityInterface $object = null): array
+  {
+    if($this->closureTranslateArgument)
+    {
+      $this->closureTranslateArgument->call($this, $this, $object);
+    }
+    return $this->translateArguments;
+  }
+
+  /**
+   * @param \Closure $closureTranslateArgument
+   *
+   * @return Fieldset
+   */
+  public function setClosureTranslateArgument(\Closure $closureTranslateArgument): Fieldset
+  {
+    $this->closureTranslateArgument = $closureTranslateArgument;
+    return $this;
+  }
+
+  /**
+   * @param string $key
+   * @param string $value
+   *
+   * @return Fieldset
+   */
+  public function addTranslateArguments(string $key, string $value): Fieldset
+  {
+    $this->translateArguments[$key] = $value;
+    return $this;
+  }
+
+  /**
+   * @param array $translateArguments
+   *
+   * @return Fieldset
+   */
+  public function setTranslateArguments(array $translateArguments): Fieldset
+  {
+    $this->translateArguments = $translateArguments;
+    return $this;
+  }
+
+  /**
    * @return MapperElementInterface|FormMapper
    */
   public function end(): MapperElementInterface
   {
     return $this->parent;
+  }
+
+  /**
+   * @return bool
+   */
+  public function isCollapse(): bool
+  {
+    return $this->collapse;
+  }
+
+  /**
+   * @param bool $collapse
+   *
+   * @return Fieldset
+   */
+  public function setCollapse(bool $collapse): Fieldset
+  {
+    $this->collapse = $collapse;
+    return $this;
   }
 
 }