+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+### Version 3.1.0 (2022-09-17)
+* Changed templatePath option in field
+* Added autocomplete to off if not used for checkbox, select, radio
+
 ### Version 3.0.0 (2021-02-10)
 * Created new repository and new branch in to Github
 * Changed the project License to the MIT